Support modifying property and field values for simple types
Can now update strings and ints for object properties and fields..

Top level objects added to the Object Inspector are marked as
read-only. Since value types cannot be updated, they do not update
the original value, the objects from the REPL cannot be updated
if they are strings or ints, etc.