Customise Object Inspector context menu
Support deleting objects being inspected.
Disable the Add Watch context menu since this would try to
add the expression to the Watch pad.

To override the default behaviour of the MacObjectValueTreeView
a new ObjectInspectorTreeView class, which derives from the
MacObjectValueTreeView, has been added. This new tree view class
defines the update handlers for the delete and add watch menu
items and since the CommandManager uses reflection to find these
methods the base class methods are not used.

Note that the delete menu is always enabled even when it is not
possible to delete an object, such as if a child node is selected.