Display value tuple element names
If a value tuple has named elements they are now displayed. This
does not work for value tuples created directly in the REPL but
does work for objects that define them as properties or fields.
The REPL does not work since the names are stored as attributes
on the field or property or variable itself and not on the
object that is passed to the Object Inspector.

  public (double Sum, int Count) ValueTupleTest { get; set; }