Disable using command
Command is not picking up new usings added. The command is using
the ScriptOptions from the CSharpKernel but this only has the usings
for the initial configuration.

The information might be available from the CSharpKernel's
ScriptState but that requires the HandleAsync method to be used
which does not support cancellation.