Focus text editor when pad gets focus
Only works when the pad is docked. When the pad is in autohide
mode then the first time the pad is opened the pad does not get
the focus when opened from the View menu. Second use of the View menu
when the pad is open in autohide mode does give the text editor the
focus. This seems to be a VS Mac IDE problem.